Option for separate yggdrasil, jotunheimr deployments.

+++ niu_heimar/cleanup_docker.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
-docker-compose down
+if [ $# -lt 1 ]; then
+  echo "Usage: ./cleanup_docker.sh <config> <rmi>"
+  echo ""
+  echo "Example: ./cleanup_docker.sh niu_heimar.yml"
+  echo ""
+  exit
+fi
+
+docker-compose -f $1 down
 docker system prune -f
 docker container rm -f $(docker container ls -aq)
 
-commands=$1
+commands=$2
 if [[ $commands == *"rmi"* ]]; then
   docker image rm -f $(docker image ls -aq)
 fi
